requirements.txt
# Base 
pytz==2019.3
Pillow==6.2.1
python-slugify==1.2.6
# Django
Django==2.1.5
# Environment
django-environ==0.4.5

# REST Framework
djangorestframework==3.10.3

# Rest password Reset
django-rest-passwordreset==1.0.0

# Pandas
pandas==0.25.3   